user route and auth

+++ routers/graphsRouter.js
@@ -2,18 +2,8 @@ const router = require('express').Router()
 
 const Graphs = require('../models/Graphs')
 
-// router.get('/', (req, res) => {
-//     Graphs.find({}).exec(function(err, r){
-//         if(!err){
-//             res.status(201).json(r)
-//         }else{
-//             res.status(500).json('Error occured')
-//         }
-//     })
-// })
-
 router.get('/', (req, res ) => {
-    const graphs = Graphs.find({})
+    const graphs = Graphs.find({}).where({user_id: req.user.user_id})
     .then(data => {
         res.status(200).json(data)
     })
@@ -22,6 +12,7 @@ router.get('/', (req, res ) => {
 
 router.post('/', (req, res) => {
     const body = req.body
+    body.user_id = req.user.id
     const graph = new Graphs(body)
     graph.save()
     .then(data => {
@@ -35,14 +26,14 @@ router.put('/:id', (req, res) => {
     let graph = req.body
     console.log('ID ', id)
     console.log('GRAPH ', graph)
-    Graphs.updateMany({_id: id}, graph)
+    Graphs.updateMany({_id: id}, graph).where({user_id: req.user.user_id})
         .then(data => res.status(200).json(data))
         .catch(err => res.status(500).json('Error occured'))
 })
 
 router.delete('/:id', (req, res) => {
     let id = req.params.id
-    Graphs.deleteOne({_id: id})
+    Graphs.deleteOne({_id: id}).where({user_id: req.user.user_id})
         .then(data => res.status(200).json(data))
         .catch(data => res.status(500).json('Error occured'))
 })
